Add proxy integration test
To be able to run the integration test, this commit also refactors the
project to have a lib.rs file. It also refactors forward_proxy and
reverse_proxy functions into slightly smaller functions that can be
called from the test while specifying custom socket options and not
instantiating a tokio runtime.